--- a/internal/lookoutv2/schema/migrations/012_add_job_submitted_index.sql
+++ b/internal/lookoutv2/schema/migrations/012_add_job_submitted_index.sql
@@ -1,0 +1,1 @@
+CREATE INDEX CONCURRENTLY IF NOT EXISTS idx_job_submitted ON job (submitted DESC);